# docs/swagger.yaml
basePath: /
definitions:
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isEmailVerified:
        type: boolean
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
  response.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        description: Can be string or map of errors
    type: object
  services.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - user
        - admin
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
  services.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  services.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    type: object
  utils.PaginationResult:
    properties:
      limit:
        description: Interface because it can be "all" or int
      page:
        type: integer
      results: {}
      totalPages:
        type: integer
      totalResults:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: A Fullstack starter kit using Go (net/http), GORM, and JWT. Matches
    PHP Native features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Starter Kit Fullstack Go Native
  version: "1.0"
paths:
  /v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send an email with a password reset link
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Request password reset
      tags:
      - Auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Login user
      tags:
      - Auth
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout by invalidating the refresh token
      parameters:
      - description: Logout Request
        in: body
        name: request
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Logout user
      tags:
      - Auth
  /v1/auth/refresh-tokens:
    post:
      consumes:
      - application/json
      description: Get new access and refresh tokens using a valid refresh token
      parameters:
      - description: Refresh Request
        in: body
        name: request
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Refresh auth tokens
      tags:
      - Auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account and return tokens
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Register a new user
      tags:
      - Auth
  /v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using a valid token
      parameters:
      - description: Reset Token
        in: query
        name: token
        required: true
        type: string
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Reset password
      tags:
      - Auth
  /v1/users:
    get:
      consumes:
      - application/json
      description: Get users with pagination, sorting, and filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page (-1 for all)
        in: query
        name: limit
        type: integer
      - description: Sort by field (e.g. name:asc)
        in: query
        name: sortBy
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Search scope (name, email, role, id)
        in: query
        name: scope
        type: string
      - description: Filter by role
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginationResult'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a user with specific role
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new user (Admin)
      tags:
      - Users
  /v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a specific user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"